# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "User Management API with Profile support",
        "title": "AvantPro Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/notification-preferences": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve notification preferences for the authenticated user. Creates defaults if none exist.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification-preferences"
                ],
                "summary": "Get user notification preferences",
                "responses": {
                    "200": {
                        "description": "Success response with preferences",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Bulk update notification preferences for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification-preferences"
                ],
                "summary": "Update user notification preferences",
                "parameters": [
                    {
                        "description": "Bulk update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NotificationPreferenceBulkUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response with updated preferences",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/notification-preferences/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all available notification event types with descriptions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification-preferences"
                ],
                "summary": "Get available notification events",
                "responses": {
                    "200": {
                        "description": "Success response with available events",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/notification-preferences/reset": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reset all notification preferences for the authenticated user to default values",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification-preferences"
                ],
                "summary": "Reset preferences to defaults",
                "responses": {
                    "200": {
                        "description": "Success response with default preferences",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/notification-preferences/test": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a test notification for the authenticated user to verify settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification-preferences"
                ],
                "summary": "Generate test notification",
                "parameters": [
                    {
                        "description": "Test notification request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TestNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success response with test notification",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/notification-preferences/{event}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a specific notification preference for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification-preferences"
                ],
                "summary": "Update single notification preference",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification event type",
                        "name": "event",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NotificationPreferenceUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response with updated preference",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Preference not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.NotificationEvent": {
            "type": "string",
            "enum": [
                "member_joined",
                "member_left",
                "member_role_changed",
                "invitation_sent",
                "invitation_accepted",
                "invitation_expired",
                "organization_update"
            ],
            "x-enum-varnames": [
                "NotificationEventMemberJoined",
                "NotificationEventMemberLeft",
                "NotificationEventMemberRoleChanged",
                "NotificationEventInvitationSent",
                "NotificationEventInvitationAccepted",
                "NotificationEventInvitationExpired",
                "NotificationEventOrganizationUpdate"
            ]
        },
        "models.NotificationPreferenceBulkItem": {
            "description": "Single notification preference item for bulk update",
            "type": "object",
            "required": [
                "event"
            ],
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "event": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.NotificationEvent"
                        }
                    ],
                    "example": "member_joined"
                }
            }
        },
        "models.NotificationPreferenceBulkUpdateRequest": {
            "description": "Bulk update notification preferences request",
            "type": "object",
            "required": [
                "preferences"
            ],
            "properties": {
                "preferences": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.NotificationPreferenceBulkItem"
                    }
                }
            }
        },
        "models.NotificationPreferenceUpdateRequest": {
            "description": "Update notification preferences request",
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.NotificationType": {
            "type": "string",
            "enum": [
                "info",
                "success",
                "warning",
                "error"
            ],
            "x-enum-varnames": [
                "NotificationTypeInfo",
                "NotificationTypeSuccess",
                "NotificationTypeWarning",
                "NotificationTypeError"
            ]
        },
...
        "models.TestNotificationRequest": {
            "description": "Test notification request",
            "type": "object",
            "required": [
                "message",
                "organization_id",
                "title",
                "type"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1,
                    "example": "This is a test notification to verify your settings"
                },
                "organization_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440002"
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1,
                    "example": "Test Notification"
                },
                "type": {
                    "enum": [
                        "info",
                        "success",
                        "warning",
                        "error"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.NotificationType"
                        }
                    ],
                    "example": "info"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}